🐛 fix: Gemma の stopSequence 根拠コメントを事実に合わせる (app Swift)
`ModelRegistry.gemma4E2B.stopSequence` は `"<|im_end|>"` だが、この文字列は
Gemma 4 の語彙に存在しない。挙動に影響はない（終端は EOG が担う）が、根拠
コメントが「Gemma と Qwen 3 は両方 `<|im_end|>` を使う」と偽の事実を断言して
いた。値は据え置き、コメントのみを訂正する (#1417 の選択肢1)。

出荷 GGUF のヘッダを直接パースして測定:
- Gemma 4 (`gemma-4-E2B-it-Q4_K_M.gguf`): 語彙 262,144 件に `<|im_end|>` は
  不在。ターンマーカーは `<|turn>` (105) / `<turn|>` (106)、いずれも
  `token_type=3 (CONTROL)`、`eos_token_id=106`
- Qwen 3 (`Qwen3-4B-Q4_K_M.gguf`): `<|im_end|>` = 151645 = `eos_token_id`、
  同じく CONTROL

コードを読んで確定した機構: `decodePiece` / `decodePieceRaw` は
`llama_token_to_piece(..., special: false)` を渡すので CONTROL トークンは
空文字にデコードされ、`nextContentTokenOrStop` は EOG で piece を追記する
前に nil を返す。したがって部分文字列の停止パスは、モデルがセンチネルを
**平文として綴った**場合にしか到達しえない — 両モデル共通で、Gemma 固有の
話ではない。`JSONResponseParser.truncateAtChatTemplateToken` が存在すること
自体が、その平文形が観測される証拠になっている。

「Gemma では決してマッチしない」とは書いていない。平文経路は両モデルとも
到達可能なまま。

- `ModelRegistry.swift`: 値そのものの場所に「意図的に据え置き」の根拠を追記
  （ここには従来コメントが一切無く、判断が読めなかった）
- `GBNFGrammarBuilder.swift`: 例示のみ差し替え。印字可能 ASCII 範囲という
  荷重主張は真であり、`special: false` 機構は llama.cpp 内部の grammar-accept
  パスには届かないので、そちらへ機構の説明は持ち込まない
- `LlamaCppIntegrationTests.swift`: `generationStopsAtImEnd` は Gemma に対して
  空虚に緑（語彙に無い文字列は構築されえない）。実態に合わせて改名し、
  何を守っていて何を守っていないかを明記
- 同ファイルが `file_length` 400 行ちょうどだったため、コメント追加で超過。
  `.claude/rules/testing.md` 公認の兄弟 extension 分割で
  `LlamaCppIntegrationTests+Streaming.swift` を新設（新 `@Suite` にはしない —
  並列実行で共有 GGUF / Metal デバイスを取り合うため）

挙動の穴（切り詰めと診断ログの ChatML 決め打ち）は #1422 に分離。

Refs #1417
